fix/feat: страница 404, Основная страница
