+++ src/components/App/App.jsx
@@ -1,28 +1,12 @@
 import React from 'react';
 import appStyles from './App.module.css';
 import AppHeader from '../AppHeader/AppHeader';
-import BurgerIngredients from '../BurgerIngredients/BurgerIngredients';
-import BurgerConstructor from '../BurgerConstructor/BurgerConstructor';
-import Modal from '../Modal/Modal';
-import ModalIngredient from '../ModalIngredient/ModalIngredient';
-import ModalOrder from '../ModalOrder/ModalOrder';
-//ИМПОРТЫ ДЛЯ DnD___________________________________________________________________________________
-import { DndProvider } from "react-dnd";
-import { HTML5Backend } from "react-dnd-html5-backend";
-//ИМПОРТЫ ДЛЯ РЕДАКСА___________________________________________________________________________________
-import { useDispatch, useSelector } from 'react-redux';
-import { getAllItems } from '../../services/actions/allIngredients';
-import {
-  SET_CURRENT_INGREDIENT,
-  DELETE_CURRENT_INGREDIENT,
-  SET_INGREDIENT_MODAL_VISIBLE,
-  SET_INGREDIENT_MODAL_INVISIBLE
-} from '../../services/actions/currentIngredient';
-import {
-  DELETE_ORDER_NUMBER,
-  SET_ORDER_MODAL_INVISIBLE,
-  placeOrder
-} from '../../services/actions/order';
+import MainPage from '../../pages/MainPage/MainPage';
+import LoginPage from '../../pages/LoginPage/LoginPage';
+
+//ИМПОРТЫ ДЛЯ РОУТИНГА___________________________________________________________________________________
+import { BrowserRouter as Router, Redirect, Route, Switch, useHistory, useLocation } from 'react-router-dom';
+import { NotFoundPage } from '../../pages/NotFoundPage/NotFoundPage';
 
 // фильтр ингредиентов по типу
 export const filterIngredients = (array, type) => {
@@ -30,85 +14,25 @@ export const filterIngredients = (array, type) => {
 }
 
 function App() {
-  const dispatch = useDispatch();
-  const { ingredientModalVisibility } = useSelector(
-    state => state.currentIngredient
-  );
-  const { orderNumber, orderError, orderModalVisibility } = useSelector(
-    state => state.order
-  );
 
-  // открытие модалки с ингредиентом
-  const handleOpenIngredientModal = (currentIngredient) => {
-    dispatch({
-      type: SET_CURRENT_INGREDIENT,
-      item: currentIngredient
-    })
-    dispatch({
-      type: SET_INGREDIENT_MODAL_VISIBLE,
-    })
-  }
-
-  React.useEffect(() => {
-    // Вызов экшена для получения всех ингредиентов от сервера
-    dispatch(getAllItems())
-  }, [])
-
-  // открытие модалки с заказом
-  const handleOpenOrderModal = (info) => {
-    dispatch(placeOrder(info, orderError))
-  }
-
-  // закрытие модалки с ингредиентом
-  const handleCloseIngredientModal = () => {
-    dispatch({
-      type: SET_INGREDIENT_MODAL_INVISIBLE,
-    })
-    dispatch({
-      type: DELETE_CURRENT_INGREDIENT
-    })
-  }
-
-  // закрытие модалки с заказом
-  const handleCloseOrderModal = () => {
-    dispatch({
-      type: SET_ORDER_MODAL_INVISIBLE
-    })
-    dispatch({
-      type: DELETE_ORDER_NUMBER
-    })
-  }
 
   return (
     <div id="app" className={appStyles.App}>
       <AppHeader />
-      <div className={appStyles.sectionContainer}>
-        <DndProvider backend={HTML5Backend}>
-          <BurgerIngredients
-            openModal={handleOpenIngredientModal}
-          />
-          <BurgerConstructor
-            openModal={handleOpenOrderModal}
-          />
-        </DndProvider>
-      </div>
+      <Router>
+        <Switch>
+        <Route path="/" exact={true}>
+            <MainPage />
+          </Route>
+          <Route path="/login" exact={true}>
+            <LoginPage />
+          </Route>
+          <Route path="*">
+            <NotFoundPage />
+          </Route>
 
-      <Modal
-        title="Детали ингредиента"
-        isModalVisible={ingredientModalVisibility}
-        closeModal={handleCloseIngredientModal}
-      >
-        <ModalIngredient />
-      </Modal>
-
-      <Modal
-        isModalVisible={orderModalVisibility}
-        closeModal={handleCloseOrderModal}
-      >
-        <ModalOrder
-          orderNumber={orderNumber}
-        />
-      </Modal>
+        </Switch>
+      </Router>
     </div>
   );
 }
