+++ src/components/App/App.module.css
@@ -4,9 +4,3 @@
   text-align: center;
   background: #131316;
 }
-
-.sectionContainer {
-  margin: 0 auto;
-  display: flex;
-  flex-direction: row;
-}
