+++ src/pages/LoginPage/LoginPage.jsx
@@ -1,0 +1,10 @@
+import styles from './LoginPage.module.css';
+import { BurgerIcon, ListIcon, Logo, ProfileIcon } from '@ya.praktikum/react-developer-burger-ui-components';
+
+const LoginPage = () => {
+  return (
+    <></>
+  )
+}
+
+export default LoginPage;
