+++ src/pages/MainPage/MainPage.module.css
@@ -1,0 +1,5 @@
+.sectionContainer {
+  margin: 0 auto;
+  display: flex;
+  flex-direction: row;
+}
