+++ src/pages/NotFoundPage/NotFoundPage.module.css
@@ -1,0 +1,32 @@
+.section {
+  width: 100%;
+  height: 800px;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+.header {
+  font-weight: bold;
+  font-size: 80px;
+  line-height: 80px;
+}
+
+.text {
+  font-size: 30px;
+  line-height: 30px;
+}
+
+.link {
+  width: 350px;
+  height: 40px;
+  padding-top: 20px;
+  text-decoration: none;
+  color: #8585ad;
+  text-align: center;
+  border-radius: 30px;
+}
+
+.link:hover {
+  background: #8585ad2a;
+}
